Add files via upload
The creation of the Anjali E-Commerce Sales Dashboard marks not only a significant milestone in the realm of data analytics but also a profound journey of learning and personal growth. Through the development of this project, i had not only gained expertise in Power BI but had also honed my skills in data visualization, business intelligence, and decision-making.